% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions_automatisation_imports.R
\name{adzipComplet}
\alias{adzipComplet}
\title{Généralisation de la fonction adzip de pmeasyr pour l'ensemble d'une archive in et out
avec contrôle du type de fichier et création des fichiers manquants}
\usage{
adzipComplet( zfichiers, ext_to_import )
}
\arguments{
\item{zfichiers}{vecteur, nom des fichiers zippés}

\item{ext_to_import}{vecteur, extention des fichiers a importer}
}
\value{
aucun, dézip des archives en fonction du type de fichier voulus et création de fichier vide pour les types de fichier manquant
}
\description{
Généralisation de la fonction adzip de pmeasyr pour l'ensemble d'une archive in et out
avec contrôle du type de fichier et création des fichiers manquants
}
\examples{
\dontrun{

   adzipComplet( zfichiers )
   
}

}
