% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions_automatisation_imports.R
\name{adzipRemonteee}
\alias{adzipRemonteee}
\title{Généralisation de la fonction adzip de pmeasyr pour l'ensemble d'une remontée (avec plusieurs sites)}
\usage{
adzipRemonteee( p, ext_to_import )
}
\arguments{
\item{p}{noyau pmeasyr}

\item{ext_to_import}{extension des fichiers à lire et importer}

\item{fichiers_genrsa}{liste des fichiers contenus dans le dossier archive}
}
\value{
aucun, dézip des archives en fonction du type de fichier voulus et création de fichier vide pour les types de fichier manquant
}
\description{
Généralisation de la fonction adzip de pmeasyr pour l'ensemble d'une remontée (avec plusieurs sites)
}
\examples{
\dontrun{

   adzipRemonteee( p )
   
}

}
