% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions_automatisation_imports.R
\name{analyse_dossier_remontees}
\alias{analyse_dossier_remontees}
\title{Scan le dossier de sauvegarde des fichiers en sortie de GENRSA, les analyse et prépare un tableau de synthèse des remontées disponibles}
\usage{
analyse_dossier_remontees( )
}
\arguments{
\item{path}{chemin du dossier dans lequel se trouve les fichiers zipper de remontée}

\item{ext}{extentinon à intégrer}
}
\value{
tibble, tableau de synthèse des remontées disponibles
}
\description{
Scan le dossier de sauvegarde des fichiers en sortie de GENRSA, les analyse et prépare un tableau de synthèse des remontées disponibles
}
\examples{
\dontrun{

   analyse_dossier_remontees( )
   
}

}
