% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions_automatisation_imports.R
\name{analyse_fichiers_remontees}
\alias{analyse_fichiers_remontees}
\title{Lecture des fichiers en sortie de scan_path pour vérification de leur contenu et préparation aux imports}
\usage{
analyse_fichiers_remontees(fichiers_genrsa)
}
\arguments{
\item{fichiers_genrsa}{tibble en sortie de scan_path}

\item{recent}{logical, si TRUE considère par défaut que le .RData du dernier mois est manquant}
}
\value{
tibble avec liste des fichiers par remontée
}
\description{
Lecture des fichiers en sortie de scan_path pour vérification de leur contenu et préparation aux imports
}
