% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions_complementaires.R
\name{attribution_statut_nx_patient}
\alias{attribution_statut_nx_patient}
\title{Attribution du statut nouveau patient pour les objets de type rsa (utilisé pour la cancérologie)
Les rsa de cancerologie sont préalablerment selectionnés en utilisant les fonctions \code{\link{selection_cancer_diag}} et
\code{\link{selection_cancer_pat}}. Les objets en sortie de ces deux fonctions sont concaténés avec un objet rsa et mis en entrée de}
\usage{
attribution_statut_nx_patient(df)
}
\arguments{
\item{df}{un tibble de type rsa concaténé avec un tibble diagnostic en sortie de \code{\link{selection_cancer_pat}}}
}
\value{
un tibble de type rsa
}
\description{
Attribution du statut nouveau patient pour les objets de type rsa (utilisé pour la cancérologie)
Les rsa de cancerologie sont préalablerment selectionnés en utilisant les fonctions \code{\link{selection_cancer_diag}} et
\code{\link{selection_cancer_pat}}. Les objets en sortie de ces deux fonctions sont concaténés avec un objet rsa et mis en entrée de
}
\examples{

}
