% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions_automatisation_imports.R
\name{attribution_type_M4}
\alias{attribution_type_M4}
\title{Attribution du type de chirurgie M4 pour les objets de type rsa
Les rsa de cancerologie sont préalablerment selectionnés en utilisant les fonctions \code{\link{selection_cancer_diag}} et
\code{\link{selection_cancer_pat}}. Les objets en sortie de ces deux fonctions sont concaténés avec un objet rsa et mis en entrée de
la fonction pour attribution du type M4 de cancer}
\usage{
attribution_type_M4(df)
}
\arguments{
\item{df}{un tibble de type rsa concaténé avec un tibble diagnostic en sortie de \code{\link{selection_cancer_pat}}}
}
\value{
un tibble de type rsa
}
\description{
Attribution du type de chirurgie M4 pour les objets de type rsa
Les rsa de cancerologie sont préalablerment selectionnés en utilisant les fonctions \code{\link{selection_cancer_diag}} et
\code{\link{selection_cancer_pat}}. Les objets en sortie de ces deux fonctions sont concaténés avec un objet rsa et mis en entrée de
la fonction pour attribution du type M4 de cancer
}
\examples{

}
