% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions_activite_donnees_pluri_annuelles.R
\name{diff_sej_recettes_v2}
\alias{diff_sej_recettes_v2}
\title{Calcul de la part liée à la variation de la répartition dans les niveaux de sévérité
dans l'évolution des recettes T2A entre 2 années.}
\usage{
diff_sej_recettes_v2(df, pivot)
}
\arguments{
\item{df}{tibble, contenant l'ensemble des données nécessaires au calcul}

\item{pivot}{string, le niveau de structure pour lequel on souhaite faire le calcul}
}
\description{
Calcul de la part liée à la variation de la répartition dans les niveaux de sévérité
dans l'évolution des recettes T2A entre 2 années.
}
