% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions_complementaires.R
\name{diff_tb}
\alias{diff_tb}
\title{Calcul de la différence et du pourcentage de différence
 entre les deux dernières colonnes d'un tableau de bord}
\usage{
diff_tb(t)
}
\arguments{
\item{t}{tableau résultats de table/tapply un tableau de bord}
}
\value{
tableau modifié
}
\description{
Calcul de la différence et du pourcentage de différence
 entre les deux dernières colonnes d'un tableau de bord
}
