% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onload.R
\name{.onLoad}
\alias{.onLoad}
\title{Actions to do when package is called}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{libname}{}

\item{pkgname}{}
}
\description{
Actions to do when package is called
}
