% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions_activite_donnees_pluri_annuelles.R
\name{get_activite_recettes}
\alias{get_activite_recettes}
\title{Tableau de bord général d'activité en recettes, et séparant HDJ/HC}
\usage{
get_activite_recettes(df, structure)
}
\arguments{
\item{df}{un tableau de données de type séjours rum/rsa}
}
\value{
tableau de bord d'activité général en recettes
}
\description{
Tableau de bord général d'activité en recettes, et séparant HDJ/HC
}
