% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions_activite_donnees_pluri_annuelles.R
\name{get_activite_sejours}
\alias{get_activite_sejours}
\title{Tableau de bord général d'activité avec un compte des séjours, et séparant HDJ/HC}
\usage{
get_activite_sejours(df, structure)
}
\arguments{
\item{df}{un tableau de données de type séjours rum/rsa}
}
\value{
tableau de bord d'activité général en séjours
}
\description{
Tableau de bord général d'activité avec un compte des séjours, et séparant HDJ/HC
}
