% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions_complementaires.R
\name{get_cmd}
\alias{get_cmd}
\title{cmd et libellé
retourne un tibble avec la liste des CMD et leur libellé}
\usage{
get_cmd()
}
\value{
tibble cmd
}
\description{
cmd et libellé
retourne un tibble avec la liste des CMD et leur libellé
}
\examples{
get_cmd()
}
