% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions_complementaires.R
\name{get_data}
\alias{get_data}
\title{Selection de données en fonction de critère temporels et géographique (lieu d'hopsitalisation)
 Applicable sur des objets en sortie de pmeasyr}
\usage{
get_data(DF, ref = "ansor", m, a, val = NULL, niveau = NULL,
  opt = T)
}
\arguments{
\item{ref}{chaine de caractères, la variable de référence qui sera utilisée pour la selection temporel
- si ref = ansor : les variables pmeasyr ansor et moisor sont utilisé pour la sélection avec les paramètres m et a
par exemple pour des objets de type rum ou rsa}

\item{m}{vecteur numérique, un vecteur de nombre entiers indiquant les mois à selectionner}

\item{a}{vecteur numérique, un vecteur de nombre entiers indiquant les années à selectionner}

\item{val}{chaine de caractères, variable géographique qui sera choisie pour la selection (ex : service, pole,...).
Si NULL pas de critères géographiques pour la sélection}

\item{niveau}{chaine de caractères, valeur de la variable géographique à sélectionner}

\item{df}{tibble, un tibble en sortie de pmeasyr}

\item{regles}{logique, utilisation des règles locales T}
}
\value{
tibble, un tibble contenant la sélection selon les critères indiqués
}
\description{
Selection de données en fonction de critère temporels et géographique (lieu d'hopsitalisation)
 Applicable sur des objets en sortie de pmeasyr
}
