% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions_complementaires.R
\name{get_diff}
\alias{get_diff}
\title{Calcul de la différence et du pourcentage
des deux dernières colonnes d'un tableau de bord}
\usage{
get_diff(df)
}
\arguments{
\item{df}{}
}
\description{
Calcul de la différence et du pourcentage
des deux dernières colonnes d'un tableau de bord
}
