% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions_automatisation_imports.R
\name{get_path_remontee}
\alias{get_path_remontee}
\title{Récupère le dossier dans lequel se trouve les archives pmsi d'une remontée}
\usage{
imco( p, tarifsante, save, persist,  )
}
\arguments{
\item{p}{un noyau pmeasyr}
}
\value{
chemin complet du dossier dans quel se trouve l'ariche pmsi
}
\description{
Récupère le dossier dans lequel se trouve les archives pmsi d'une remontée
}
\examples{
\dontrun{

   get_path_remontee( p ) -> path
}

}
