% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions_complementaires.R
\name{get_racines}
\alias{get_racines}
\title{racines de ghm et libellé
retourne un tibble avec la liste des CMD et leur libellé}
\usage{
get_racines()
}
\value{
tibble racines
}
\description{
racines de ghm et libellé
retourne un tibble avec la liste des CMD et leur libellé
}
\examples{
get_racines()
}
