% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions_activite_donnees_pluri_annuelles.R
\name{get_tdb}
\alias{get_tdb}
\title{Création d'un tableau indicateurs x pivot, comprenant une liste d'inficateurs}
\usage{
get_tdb(df, indicateurs, pivot = "ansor")
}
\arguments{
\item{df}{un tibble de type séjours}

\item{indicateurs}{vecteur, liste des indicateurs à calculer}

\item{pivot}{string, la variable pivot a utiliser dans df pour réaliser le tableau croisé}
}
\value{
tableau de bord
}
\description{
Création d'un tableau indicateurs x pivot, comprenant une liste d'inficateurs
}
