% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions_automatisation_imports.R
\name{imco}
\alias{imco}
\title{Import des principaux fichiers de remontées éavec pmeasyr et valorisation des séjours et résumés}
\usage{
imco( p )
}
\arguments{
\item{p}{un noyau pmeasyr}

\item{tarifsante}{si TRUE utilisation des tarifs de l'année antérieure}

\item{save}{si TRUE sauvegarde un fichier .RData dans le répertoire des données de remontée}

\item{persist}{si TRUE renvoie les données}
}
\value{
liste avec principaux fichiers de remontées, et enregistrement sur le disque si besoin
}
\description{
Import des principaux fichiers de remontées éavec pmeasyr et valorisation des séjours et résumés
}
\examples{
\dontrun{

   imco( p ) -> data_mco
}

}
