% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions_automatisation_imports.R
\name{import_remontees}
\alias{import_remontees}
\title{Création des fichiers .RData contenant l'ensemble des données d'une remontée à partir d'un noyau pmeasyr}
\usage{
import_remontees( p )
}
\arguments{
\item{p}{}
}
\description{
Création des fichiers .RData contenant l'ensemble des données d'une remontée à partir d'un noyau pmeasyr
}
\examples{
\dontrun{

   import_remontees( p )
   
}

}
