% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions_complementaires.R
\name{IP_SEJOUR}
\alias{IP_SEJOUR}
\title{Calcul de l'index de performance à partir d'un objet de type séjours rum/rsa
 Attention pas de dédoublonnage}
\usage{
IP_SEJOUR(df)
}
\arguments{
\item{df}{}
}
\description{
Calcul de l'index de performance à partir d'un objet de type séjours rum/rsa
 Attention pas de dédoublonnage
}
