% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions_complementaires.R
\name{IP_SERVICE}
\alias{IP_SERVICE}
\title{Calcul de l'index de performance SERVICE à partir d'un objet de type séjours rum
 L'objet en entrée doit avoir une variable service}
\usage{
IP_SERVICE(df)
}
\arguments{
\item{df}{objet de type séjours (rum/rsa)}
}
\value{
double, index de performance
}
\description{
Calcul de l'index de performance SERVICE à partir d'un objet de type séjours rum
 L'objet en entrée doit avoir une variable service
}
