% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions_automatisation_imports.R
\name{load_all}
\alias{load_all}
\title{charge en mémoire les objets définitifs rum,rum_v,rsa,rsa_v,diagnostics, actes, vano}
\usage{
load_all(remontees_sel)
}
\arguments{
\item{remontees_sel}{tibble de type analyse_remontee avec l'ensemble des année à charger}
}
\description{
charge en mémoire les objets définitifs rum,rum_v,rsa,rsa_v,diagnostics, actes, vano
}
