% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions_automatisation_imports.R
\name{load_dmi}
\alias{load_dmi}
\title{Chargement des fichiers de DMI en entrée et sortie de GENRSA avec pmeasyr}
\usage{
load_dmi( remontees_sel )
}
\arguments{
\item{remontees_sel}{}
}
\description{
Chargement des fichiers de DMI en entrée et sortie de GENRSA avec pmeasyr
}
\examples{
\dontrun{

   load_dmi( remontees_sel )
   
}

}
