% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions_automatisation_imports.R
\name{load_RData}
\alias{load_RData}
\title{charge en mémoire les objets définitifs rum,rum_v,rsa,rsa_v,diagnostics, actes, vano préparé dans le RData}
\usage{
load_RData( sel_remontees_import )
}
\arguments{
\item{sel_remontees_import}{tibble de type analyse_remontee avec l'ensemble des années à charger}

\item{extra}{logical, si TRUE importe également :
- les données non consolidée (le même mois de la remontée de l'année en cours) de l'année antérieure, 
- les données de l'année valorisées avec les tarifs de l'année antérieure}
}
\description{
charge en mémoire les objets définitifs rum,rum_v,rsa,rsa_v,diagnostics, actes, vano préparé dans le RData
}
\examples{
\dontrun{

   load_RData( sel_remontees_import )
   
}

}
