% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions_automatisation_imports.R
\name{maj_variable_RData}
\alias{maj_variable_RData}
\title{Mise à jour de la variabla RData du tableau de synthèse des remontées disponibles pour la dernière remontée disponible
force la variable RData à 0 pour que la dernière remontée disponible soit importée}
\usage{
maj_variable_RData(remontees, p = NULL)
}
\arguments{
\item{remontees}{tibble, tableau de synthèse des remontées disponibles}

\item{p}{un noyau pmeasyr}
}
\description{
Mise à jour de la variabla RData du tableau de synthèse des remontées disponibles pour la dernière remontée disponible
force la variable RData à 0 pour que la dernière remontée disponible soit importée
}
