% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions_activite_donnees_pluri_annuelles.R
\name{make_tdb}
\alias{make_tdb}
\title{Création des tableaux de bord}
\usage{
make_tdb(val, niveau, annee, mois )
}
\arguments{
\item{val}{}

\item{mois}{}
}
\value{
tableau de bord d'activité
}
\description{
Création des tableaux de bord
}
\examples{
\dontrun{

   get_tdb(df, indicateurs) -> tdb
}
}
