% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions_complementaires.R
\name{nb_unique}
\alias{nb_unique}
\title{Nombbre d'élément unique dans un vecteur}
\usage{
nb_unique(x)
}
\arguments{
\item{x}{vecteur}
}
\value{
integer, nombre d'élément unique
}
\description{
Nombbre d'élément unique dans un vecteur
}
\examples{

}
