% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions_complementaires.R
\name{options_locales}
\alias{options_locales}
\title{Mise en oeuvre des options locales pour le décompte des séjours dans les données d'activité}
\usage{
options_locales(DF, val = NULL, niveau = NULL)
}
\arguments{
\item{val}{chaine de caractères, variable géographique qui sera choisie pour la selection (ex : service, pole,...).}

\item{niveau}{chaine de caractères, variable géographique qui sera choisie pour la selection (ex : service, pole,...).}

\item{df}{tibble de type rum/rsa, utilisé entre autres en entrée de \code{\link{get_data}}}
}
\value{
tibble de type rum/rsa, le tibble local auquel a été ajoutée.
}
\description{
La fonction ajoute une variable doublon ( TRUE / FALSE ) à un objet de type rum/rsa
qui sera utilisée dans les tableaux de bord pour dédoublonner les séjours au moment de leur décompte
}
\examples{


}
