% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions_complementaires.R
\name{order_by_structure}
\alias{order_by_structure}
\title{Utilisé pour la tableaux de bord généraux
Reformatage d'objets en sortie de table/tapply selon l'ordonnancement décidé dans le fichier de structure
 Fusion des tableaux}
\usage{
order_by_structure(tdb)
}
\arguments{
\item{tdb}{list d'objets en sortie de table/tapply avec 3 dimenssions

       - Nom du niveau de structure (ex nom de service ou de pole)
       - Variable de temporalité (par exemple l'année)
       - C / P}
}
\value{
concaténation et restructuration des tableaux de données
}
\description{
Utilisé pour la tableaux de bord généraux
Reformatage d'objets en sortie de table/tapply selon l'ordonnancement décidé dans le fichier de structure
 Fusion des tableaux
}
