% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions_complementaires.R
\name{prep_string}
\alias{prep_string}
\title{Suppression des caractères spéciaux et accentués}
\usage{
prep_string(string)
}
\arguments{
\item{string}{chaine de caractères}
}
\description{
Suppression des caractères spéciaux et accentués
}
\examples{
 
 
}
