% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions_complementaires.R
\name{prep_tb}
\alias{prep_tb}
\title{Fonction de tableaux de bord
Mise en forme du tableau -> remplacement des lignes vides par '-'}
\usage{
prep_tb(tab)
}
\arguments{
\item{tab}{tableau résultats de table/tapply un tableau de bord}
}
\value{
tableau modifié
}
\description{
Fonction de tableaux de bord
Mise en forme du tableau -> remplacement des lignes vides par '-'
}
\examples{

}
