% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions_automatisation_imports.R
\name{scan_path}
\alias{scan_path}
\title{Récupération de la liste de l'ensemble des fichiers de remontée contenu dans un dossier}
\usage{
scan_path(path = getOption("dimRactivite.path"),
  ext = getOption("dimRactivite.extensions"))
}
\arguments{
\item{path}{racine du dossier contenant les archives}
}
\value{
liste des fichiers d'archives disponibles
}
\description{
Récupération de la liste de l'ensemble des fichiers de remontée contenu dans un dossier
}
\examples{
fichiers_genrsa <- scan_path(getOption("dimRactivite.path"))

}
