% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions_complementaires.R
\name{selection_cancer_diag}
\alias{selection_cancer_diag}
\title{un ensemble de fonction permettant d'automatiser les imports pmeasyr et d'effectuer des opérations
de transformation et de sélection des séjours}
\usage{
selection_cancer_diag(diagnostics)
}
\arguments{
\item{df}{un tibble de type diagnostic}
}
\value{
une sélection de df pour la cancerologie
}
\description{
Warning: utilisation du package pmeasyr
Selection des diagnostics de cancerologie pour un objet pmeasyr diagnostics
}
\section{Warning}{
 utilise la liste inca_cancero
}

