% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions_automatisation_imports.R
\name{selection_cancer_pat}
\alias{selection_cancer_pat}
\title{Attribution d'un diagnostic de cancer par patient pour un objet de type diagnotics en sortie de \code{\link{selection_cancer_diag}}}
\usage{
selection_cancer_pat(df)
}
\arguments{
\item{df}{un tibble de type diagnostic en sortie de \code{\link{selection_cancer_diag}}}
}
\value{
un tibble de type diagnostic comprenant une ligne par patient selectionne comme etant le diagnostic de cancer unique pour l'analyse de l'activite
}
\description{
Attribution d'un diagnostic de cancer par patient pour un objet de type diagnotics en sortie de \code{\link{selection_cancer_diag}}
}
\section{Warning}{
 utilise lae resultat de la fonction  \code{\link{selection_cancer_diag}}
}

\examples{

}
