% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onload.R
\name{set_option}
\alias{set_option}
\title{set option from val}
\usage{
set_option(opt, val)
}
\arguments{
\item{opt}{string, name of the option}

\item{val}{string/number, value of the option}
}
\description{
set option from val
}
