% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onload.R
\name{update_options}
\alias{update_options}
\title{Chargement des options du package dimRactivité}
\usage{
update_options(file = "options.yaml")
}
\arguments{
\item{file}{fichier yaml d'options du package dimRactivité

Options à renseigner dans le fichier sont:}

\item{services_exclus}{: lise des services exclus (tels qu'il apparaissent dans le fichier structure) séparé par une virgule}

\item{gestion_doublons_rum}{: pour certain niveau de structure ou certain libellé de regroupement}
}
\description{
Chargement des options du package dimRactivité
}
\examples{
update_options()
}
