% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions_automatisation_imports.R
\name{update_remontees_dispo}
\alias{update_remontees_dispo}
\title{Title}
\usage{
update_remontees_dispo(path = getOption("dimRactivite.path"),
  ext = getOption("dimRactivite.extensions"))
}
\arguments{
\item{path}{chemin du dossier dans lequel se trouve les fichiers zipper de remontée}

\item{ext}{extentinon à intégrer}
}
\value{
tibble, tableau de synthèse des remontées disponibles
}
\description{
Title
}
