% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions_automatisation_imports.R
\name{verif_structure}
\alias{verif_structure}
\title{Vérification de l'exhaustivité du fichier structure}
\usage{
verif_structure(rum, fichier_structure)
}
\arguments{
\item{rum}{tiblle de type rum}

\item{fichier_structure}{tibble, structure}
}
\description{
Vérification de l'exhaustivité du fichier structure
}
