% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions_activite_donnees_pluri_annuelles.R
\name{get_tdb}
\alias{get_tdb}
\title{Création d'un tableau indicateurs x pivot, comprenant une liste d'indicateurs}
\usage{
get_tdb(df, indicateurs)
}
\arguments{
\item{df}{un tibble de type séjours}

\item{indicateurs}{vecteur, liste des indicateurs à calculer}

\item{pivot}{string, la variable pivot a utiliser dans df pour réaliser le tableau croisé, 
peut être soit un facteur soit une date, si une date la variable unit_pivot doit être renseignée}

\item{unit_pivot}{string, la variable pivot a utiliser dans df pour réaliser le tableau croisé}
}
\value{
dataframe contenant les statistiques demandées dans indicateurs mise en forme
}
\description{
Création d'un tableau indicateurs x pivot, comprenant une liste d'indicateurs
}
\examples{
\dontrun{

   get_tdb(df, indicateurs) -> tdb
   
}

}
