% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions_automatisation_imports.R
\name{load_RData}
\alias{load_RData}
\title{charge en mémoire les objets définitifs rum,rum_v,rsa,rsa_v,diagnostics, actes, vano préparé dans le RData}
\usage{
load_RData( remontees_sel )
}
\arguments{
\item{remontees_sel}{tibble de type analyse_remontee avec l'ensemble des années à charger}
}
\description{
charge en mémoire les objets définitifs rum,rum_v,rsa,rsa_v,diagnostics, actes, vano préparé dans le RData
}
\examples{
\dontrun{

   load_RData( remontees_sel )
   
}

}
