% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions_automatisation_imports.R
\name{save_remontees}
\alias{save_remontees}
\title{Création des fichiers .RData contenant l'ensemble des données d'une remontée à partir du tableau d'analyse des fichiers de remontée (analyse_remontee)}
\usage{
save_remontees(dossiers_remontees, fichiers_genrsa)
}
\arguments{
\item{dossiers_remontees}{tibble en sortie de  analyse_remontee}

\item{fichiers_genrsa}{tibble en sortie de  scan_path avec analyse de l'ensemble des fichiers du dossier racine}
}
\value{
ecriture sur le disque d'1 .RData de sauvegarde pour chaque remontée
}
\description{
Création des fichiers .RData contenant l'ensemble des données d'une remontée à partir du tableau d'analyse des fichiers de remontée (analyse_remontee)
}
